New Matching System
Info:
Code works on the server!

Location:
Serverscriptservice > Server

Description:
when the player sits on the chair, and if there is no one else sitting he will be defined as the owner, if there are players in the same seat as the owner and he leaves, the other players will jump from the chairs, like a forced jump.