Teleport and Fixed DataStore Player
Info:
Code Server and Client

Location:
ServerScriptService > MatchingHandler
ServerScriptService > PlayerCollision
ServerScriptService > Datastore

Description:
In this Push, I completed the player's official teleport system, as usual since the previous post, when the player sat on the bench he opened a Gui to teleport, if the owner left the bench all the other members were forced to jump from the bench. banks, now the teleportation system has been implemented, when all members have already been seated on the bench and when the owner presses the teleport button, everyone is teleported, it was just configured to teleport to a specific place, in the future the code will undergo changes for new maps, when that moment arrives I will have to redo a part of the code, I fixed the player datastore that only saved values that were created by code, I fixed the "PlayerCollisions" code, that's all for today.