New Assets and Systems
Info:
Code Server and Client

Location:
StarterGui > Client > Hud

Description:
I then added a counter/Timer so that when the player starts the game they have to wait 60 seconds, which will give time for other players to enter and participate. If the party owner leaves the game or leaves the bench, the counter will be canceled and everyone will be forced to leave the game. party, with this I took advantage of the fact that PWE is implementing the towers hotbar interface and the yen interface. I already configured the hotbar with the inventory, I also configured the Yen with the Yen interface with the animation of money going up and down.